Implement user endpoint

+++ src/controller/UserController.ts
@@ -3,14 +3,15 @@ import {NextFunction, Request, Response} from "express";
 import {User} from "../entity/User";
 
 export class UserController {
-
     private userRepository = getRepository(User);
 
     async all(request: Request, response: Response, next: NextFunction) {
         return this.userRepository.find();
     }
 
-    async one(request: Request, response: Response, next: NextFunction) {
+    // http://localhost:8080/user/61cb8827a45ace8fcf0ec1c1
+    async one(request: Request, response: Response) {
+        console.log(request.params.id)
         return this.userRepository.findOne(request.params.id);
     }
 
@@ -20,7 +21,7 @@ export class UserController {
 
     async remove(request: Request, response: Response, next: NextFunction) {
         let userToRemove = await this.userRepository.findOne(request.params.id);
-        await this.userRepository.remove(userToRemove);
+        if (userToRemove)
+          await this.userRepository.remove(userToRemove);
     }
-
 }
